requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
awsebcli==3.21.0
blessed==1.20.0
blinker==1.9.0
botocore==1.35.71
cachelib==0.13.0
cement==2.10.14
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
MarkupSafe==3.0.2
msgspec==0.19.0
openai==1.55.3
packaging==24.2
pathspec==0.10.1
pydantic==2.10.2
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
